Curso Java1

+++ src/For3.java
@@ -4,7 +4,7 @@ public class For3 {
 
 	public static void main(String[] args) { //Hallar factorial de un numero
 		
-		int resultado=1;
+		Long resultado=1L;
 		int numero=Integer.parseInt(JOptionPane.showInputDialog("Introduzca un número"));
 		
 		for (int i = numero; i > 0; i--) {
